API Stats, My Uploads, Profile NavBar
closes #27 #28 #29
API stats are now pulling appropriately. Only owner of the mode can see model wide stats. This is done using request and comparing to the model author. Otherwise if this condition fails, the owner level stats of the model is returned.
The Upload page is complete utilizing a re-usable banner for the model. Maybe try to make banner generic enough to use in other areas?